Rules.
Added a good number of rules, as well as a "maincolor" string that will
hold the value when asked what the main color of your animal is. I'll
update this in a  bit with some more rules, but as far as question I
need to talk to Ben about the format.